change ADC sample rate to 1kHz
this costs us about 9% of our CPU, but should make aliasing much less
of a problem.

+++ libraries/AP_ADC/AP_ADC_ADS7844.h
@@ -24,7 +24,7 @@ class AP_ADC_ADS7844 : public AP_ADC
 	uint16_t Ch(unsigned char ch_num);
 
 	// Read 6 sensors at once
-	uint16_t Ch6(const uint8_t *channel_numbers, uint16_t *result);
+	uint32_t Ch6(const uint8_t *channel_numbers, uint16_t *result);
 
 	private:
 };
